refactor: Fix file structure
